fix(indicator,sizing,strategy,event,fills): round every product before the sum it feeds
A journal was architecture-dependent. Go permits fusing a + b*c into a
single fused multiply-add "possibly across statements", and arm64 does while
amd64 does not, so the fused form's full-precision product produced
different last bits on the two machines. The whole-life exit price in
campaign.go was the expression the golden journal caught:

  exit_price  118.2875 on arm64, 118.28750000000001 on amd64

and with it realised_result, average_move_in_n and realised_result_in_unit_n.
This is why CI failed on amd64 while make check passed locally, and it is a
determinism defect in its own right (.greptile/rules.md): replay equivalence
and ADR 0012's "same configuration, same results" both assume a number does
not depend on the machine that computed it.

Every product feeding an addition or subtraction in internal/ is now rounded
explicitly, which is the only barrier the language guarantees — in the
producers and in the event payload validators that re-derive the same
values, so the two cannot disagree. Found by walking the AST for
add/subtract nodes with a multiply operand rather than by reading.

No value changes on amd64: the regenerated golden is byte-identical to what
an amd64 build produced before this commit, and arm64 now agrees with it.
RulesVersion is not bumped — no rule changed, and no journal that was
correct becomes unreplayable.

Verified by running the whole suite, golden included, natively on arm64 and
under GOARCH=amd64.

+++ internal/sizing/ladder.go
@@ -46,7 +46,7 @@ func NextAddLevel(previousFill, campaignN float64, direction string) (float64, e
 	if err := errors.Join(errs...); err != nil {
 		return 0, fmt.Errorf("sizing: cannot derive next add level: %w", err)
 	}
-	return previousFill + 0.5*campaignN, nil
+	return previousFill + float64(0.5*campaignN), nil
 }
 
 // AddLadder returns the whole INTENDED Add Ladder for a Campaign: firstFill,
@@ -94,7 +94,7 @@ func AddLadder(firstFill, campaignN float64, maxUnits int, direction string) ([]
 		// actual fill — the two agree bit for bit whenever every fill lands
 		// exactly on its own rung, which is the assumption this function
 		// deliberately makes (see the type's doc comment).
-		ladder[i] = ladder[i-1] + 0.5*campaignN
+		ladder[i] = ladder[i-1] + float64(0.5*campaignN)
 	}
 	return ladder, nil
 }
